[Components] [Post] moved post
